Replace WinUI3 2020 project by result of latest
