+++ GettingStartedGuides/WinUI3_GettingStarted/README.md
@@ -5,11 +5,11 @@ languages:
 page_type: sample
 products: 
   - microsoft-edge
-urlFragment: WinUI3_GettingStarted
+urlFragment: WinUI3GetStarted
 ---
 # Get started with WebView2 in WinUI 3 (Windows App SDK) apps
 
 <!-- only enough info to differentiate this sample vs the others; what is different about this sample compared to the sibling samples? -->
-This sample, **WinUI3_GettingStarted**, is the completed Visual Studio project that results from following the steps in the tutorial [Get started with WebView2 in WinUI 3 (Windows App SDK) apps](https://learn.microsoft.com/microsoft-edge/webview2/get-started/winui).
+This sample, **WinUI3GetStarted**, is the completed Visual Studio project that results from following the steps in the tutorial [Get started with WebView2 in WinUI 3 (Windows App SDK) apps](https://learn.microsoft.com/microsoft-edge/webview2/get-started/winui).
 
-![The running app from the finished WinUI 3 (Windows App SDK) tutorial](screenshots/getting-started-part-3.png)
+![The running app from the finished WinUI 3 (Windows App SDK) tutorial](screenshots/finished-app.png)
