+++ GettingStartedGuides/WinUI3_GettingStarted/WinUI3GetStarted/WinUI3GetStarted.sln
@@ -1,0 +1,43 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35327.3
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WinUI3GetStarted", "WinUI3GetStarted\WinUI3GetStarted.csproj", "{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|ARM64 = Debug|ARM64
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|ARM64 = Release|ARM64
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|ARM64.Build.0 = Debug|ARM64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|ARM64.Deploy.0 = Debug|ARM64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|x64.ActiveCfg = Debug|x64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|x64.Build.0 = Debug|x64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|x64.Deploy.0 = Debug|x64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|x86.ActiveCfg = Debug|x86
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|x86.Build.0 = Debug|x86
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Debug|x86.Deploy.0 = Debug|x86
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|ARM64.ActiveCfg = Release|ARM64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|ARM64.Build.0 = Release|ARM64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|ARM64.Deploy.0 = Release|ARM64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|x64.ActiveCfg = Release|x64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|x64.Build.0 = Release|x64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|x64.Deploy.0 = Release|x64
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|x86.ActiveCfg = Release|x86
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|x86.Build.0 = Release|x86
+		{FBF1E5FC-C14D-4C2D-88D4-6B8B55821E52}.Release|x86.Deploy.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1334E548-2468-4E8F-84AF-B2B85C674856}
+	EndGlobalSection
+EndGlobal
