+++ GettingStartedGuides/WinUI3_GettingStarted/WinUI3GetStarted/WinUI3GetStarted/MainWindow.xaml.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
+using Microsoft.Web.WebView2.Core;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Controls.Primitives;
@@ -12,11 +6,18 @@ using Microsoft.UI.Xaml.Data;
 using Microsoft.UI.Xaml.Input;
 using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
-using ABI.Microsoft.UI.Private.Controls;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices.WindowsRuntime;
+using Windows.Foundation;
+using Windows.Foundation.Collections;
 
-// The Blank Window item template is documented at https://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
+// To learn more about WinUI, the WinUI project structure,
+// and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace WinUI_Sample
+namespace WinUI3GetStarted
 {
     /// <summary>
     /// An empty window that can be used on its own or navigated to within a Frame.
@@ -27,17 +28,9 @@ namespace WinUI_Sample
         {
             this.InitializeComponent();
             MyWebView.NavigationStarting += EnsureHttps;
-            //MyWebView.WebMessageReceived += UpdateAddressBar;
-            
         }
 
-        private void UpdateAddressBar(WebView2 sender, WebView2WebMessageReceivedEventArgs args)
-        {
-            //String uri = args.Source;
-            //addressBar.Text = uri;
-        }
-
-        private void EnsureHttps(WebView2 sender, WebView2NavigationStartingEventArgs args)
+        private void EnsureHttps(WebView2 sender, CoreWebView2NavigationStartingEventArgs args)
         {
             String uri = args.Uri;
             if (!uri.StartsWith("https://"))
@@ -51,11 +44,8 @@ namespace WinUI_Sample
             }
         }
 
-
-private void myButton_Click(object sender, RoutedEventArgs e)
+        private void myButton_Click(object sender, RoutedEventArgs e)
         {
-            //myButton.Content = "Clicked";
-
             try
             {
                 Uri targetUri = new Uri(addressBar.Text);
@@ -63,9 +53,8 @@ private void myButton_Click(object sender, RoutedEventArgs e)
             }
             catch (FormatException ex)
             {
-                // Bad address.
+                // Incorrect address entered.
             }
-
         }
     }
 }
