+++ GettingStartedGuides/WinUI3_GettingStarted/WinUI3GetStarted/WinUI3GetStarted/Properties/launchSettings.json
@@ -1,0 +1,10 @@
+{
+  "profiles": {
+    "WinUI3GetStarted (Package)": {
+      "commandName": "MsixPackage"
+    },
+    "WinUI3GetStarted (Unpackaged)": {
+      "commandName": "Project"
+    }
+  }
+}
