--- GettingStartedGuides/WinUI3_GettingStarted/WinUI_Sample/WinUI_Sample.sln
@@ -1,53 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30223.230
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5}") = "WinUI_Sample (Package)", "WinUI_Sample (Package)\WinUI_Sample (Package).wapproj", "{B820339B-BF7D-4F43-826D-D1A82A763178}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WinUI_Sample", "WinUI_Sample\WinUI_Sample.csproj", "{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Debug|x64.ActiveCfg = Debug|x64
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Debug|x64.Build.0 = Debug|x64
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Debug|x64.Deploy.0 = Debug|x64
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Debug|x86.ActiveCfg = Debug|x86
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Debug|x86.Build.0 = Debug|x86
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Debug|x86.Deploy.0 = Debug|x86
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Release|Any CPU.ActiveCfg = Release|x86
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Release|x64.ActiveCfg = Release|x64
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Release|x64.Build.0 = Release|x64
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Release|x64.Deploy.0 = Release|x64
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Release|x86.ActiveCfg = Release|x86
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Release|x86.Build.0 = Release|x86
-		{B820339B-BF7D-4F43-826D-D1A82A763178}.Release|x86.Deploy.0 = Release|x86
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Debug|x64.ActiveCfg = Debug|x64
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Debug|x64.Build.0 = Debug|x64
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Debug|x86.ActiveCfg = Debug|x86
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Debug|x86.Build.0 = Debug|x86
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Release|x64.ActiveCfg = Release|x64
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Release|x64.Build.0 = Release|x64
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Release|x86.ActiveCfg = Release|x86
-		{D507D20A-6AFB-4B0F-AEF0-C5D99F3EBE08}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {AD0AABFA-D6C4-43AA-8C9D-D33713CD0977}
-	EndGlobalSection
-EndGlobal
