--- GettingStartedGuides/WinUI3_GettingStarted/WinUI_Sample/WinUI_Sample/App.xaml
@@ -1,15 +1,0 @@
-﻿<Application
-  x:Class="WinUI_Sample.App"
-  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:local="using:WinUI_Sample">
-  <Application.Resources>
-    <ResourceDictionary>
-      <ResourceDictionary.MergedDictionaries>
-        <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-        <!-- Other merged dictionaries here -->
-      </ResourceDictionary.MergedDictionaries>
-      <!-- Other app resources here -->
-    </ResourceDictionary>
-  </Application.Resources>
-</Application>
