--- GettingStartedGuides/WinUI3_GettingStarted/WinUI_Sample/WinUI_Sample/MainWindow.xaml
@@ -1,30 +1,0 @@
-﻿<Window
-  x:Class="WinUI_Sample.MainWindow"
-  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:local="using:WinUI_Sample"
-  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-  mc:Ignorable="d"
-  xmlns:controls="using:Microsoft.UI.Xaml.Controls">
-
-
-    <Grid>
-
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="Auto" />
-        </Grid.ColumnDefinitions>
-    
-        <TextBox Name="addressBar" Grid.Column="0"/>
-        <Button x:Name="myButton" Grid.Column="1" Click="myButton_Click">Go</Button>
-
-        <controls:WebView2 x:Name="MyWebView"  controls:Grid.Row="1" controls:Grid.ColumnSpan="2" Source="https://www.microsoft.com" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" />
-
-    </Grid>
-    
-</Window>
