--- GettingStartedGuides/WinUI3_GettingStarted/WinUI_Sample/WinUI_Sample/app.manifest
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1">
-  <assemblyIdentity version="1.0.0.0" name="WinUI_Sample.app"/>
-
-  <application xmlns="urn:schemas-microsoft-com:asm.v3">
-    <windowsSettings>
-      <!-- The combination of below two tags have the following effect:
-           1) Per-Monitor for >= Windows 10 Anniversary Update
-           2) System < Windows 10 Anniversary Update
-      -->
-      <dpiAware xmlns="http://schemas.microsoft.com/SMI/2005/WindowsSettings">true/PM</dpiAware>
-      <dpiAwareness xmlns="http://schemas.microsoft.com/SMI/2016/WindowsSettings">PerMonitorV2, PerMonitor</dpiAwareness>
-    </windowsSettings>
-  </application>
-</assembly>
